Extend metrics to make plottable values
By sorting the proportion of commands processed we get something
akin to a ROC curve where performance is better the closer we get the
curve to the upper-left corner.

Also moved JSON to Git (to make it plottable without getting all the
other data, and because the files are small) and pretty-printed.