Restructure movement/blitz to handle Wild Animal
Previously the "unused" object held the TargetSpaceCommand so that the
Wild Animal (or Stupid) check could look for its dice roll. Now we're
avoiding the "unused" parameter and using `peek()` then we need to
get the moves *and* the target and get them processed. BUT we need
to keep the old method as-is because it's still valid for movement.