Refactor turn handling
Automatically starting the next turn was causing problems with some
data files. It was also messy having the board responsible for some
events but not all.

The new structure includes and exception for Abandon Match so that
it guarantees that we break whatever we're doing and handle it rather
than having special case code everywhere.