Avoid invalid dodge on Leap out of tackle zones
A leap is two move commands (potentially through an occupied space)
with a Leap skill event in the log. We tracked when the last move was
a leap so we could avoid moving them on top of someone BUT we
*didn't* handle them leaping from between two players so that
the second move looked like a dodge out of a tackle zone.